tests/wpt/mozilla/tests/mozilla/bluetooth/getDescriptors/get-different-descriptors-after-reconnection.html

<!doctype html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.two_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [generic_access.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => {
        let descriptors1;
        return gattServer.getPrimaryService(generic_access.name)
        .then(service => service.getCharacteristic(device_name.name))
        .then(characteristic => characteristic.getDescriptors(number_of_digitals.name))
        .then(descriptors => descriptors1 = descriptors)
        .then(() => gattServer.disconnect())
        .then(() => gattServer.connect())
        .then(() => gattServer.getPrimaryService(generic_access.name))
        .then(service => service.getCharacteristic(device_name.name))
        .then(characteristic => characteristic.getDescriptors(number_of_digitals.name))
        .then(descriptors2 => [descriptors1, descriptors2])
    })
    .then(descriptors_arrays => {
        for (let i = 1; i < descriptors_arrays.length; i++) {
            assert_equals(descriptors_arrays[0].length, descriptors_arrays[i].length);
        }
        let base_set = new Set(descriptors_arrays.shift());
        for (let descriptors of descriptors_arrays) {
            descriptors.forEach(descriptor => assert_false(base_set.has(descriptor)));
        }
    });
}, 'Calls to getDescriptors after a disconnection should return a different object.');
</script>
